+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+
+- Removed a `print` statement that was always occurring.
+- Added a high-level `FirebaseAuth` class, and Google sign-in.
+- Added `IfFirebaseAuthDirective`.
+
 ## 0.1.0
 
 - Initial commit of `GoogleSignInComppnent`.